Isolate arrow code #237
