+++ Cargo.toml
@@ -20,8 +20,8 @@ ark-poly = { version = "0.4.0" }
 ark-serialize = { version = "0.4.0" }
 ark-std = { version = "0.4.0", default-features = false }
 arrayvec = { version = "0.7", default-features = false }
-arrow = { version = "51.0" }
-arrow-csv = { version = "51.0" }
+arrow = { version = "53.0" }
+arrow-csv = { version = "53.0" }
 bit-iter = { version = "1.1.1" }
 bigdecimal = { version = "0.4.5", default-features = false, features = ["serde"] }
 blake3 = { version = "1.3.3", default-features = false }
