+++ crates/proof-of-sql/src/arrow/base_commitment_arrow/mod.rs
@@ -1,0 +1,1 @@
+pub mod table_commitment_arrow;
