+++ crates/proof-of-sql/src/arrow/base_database_arrow/mod.rs
@@ -1,0 +1,22 @@
+pub mod column_arrow;
+#[cfg(feature = "arrow")]
+pub mod arrow_array_to_column_conversion;
+#[cfg(feature = "arrow")]
+pub use arrow_array_to_column_conversion::{ArrayRefExt, ArrowArrayToColumnConversionError};
+#[cfg(feature = "arrow")]
+mod record_batch_utility;
+#[cfg(feature = "arrow")]
+pub use record_batch_utility::ToArrow;
+#[cfg(all(test, feature = "arrow", feature = "test"))]
+mod test_accessor_utility;
+#[cfg(all(test, feature = "arrow", feature = "test"))]
+pub use test_accessor_utility::{make_random_test_accessor_data, RandomTestAccessorDescriptor};
+#[cfg(feature = "arrow")]
+mod owned_and_arrow_conversions;
+#[cfg(feature = "arrow")]
+pub use owned_and_arrow_conversions::OwnedArrowConversionError;
+#[cfg(all(test, feature = "arrow"))]
+mod owned_and_arrow_conversions_test;
+/// Contains traits for scalar <-> i256 conversions
+#[cfg(feature = "arrow")]
+pub mod scalar_and_i256_conversions;
