+++ crates/proof-of-sql/src/arrow/sql_proof_arrow/mod.rs
@@ -1,0 +1,4 @@
+pub mod query_result_arrow;
+pub mod final_round_builder_test_arrow;
+#[cfg(all(test, feature = "arrow"))]
+mod provable_query_result_test;
