+++ crates/proof-of-sql/src/lib.rs
@@ -7,6 +7,7 @@ extern crate alloc;
 
 pub mod base;
 pub mod proof_primitive;
+pub mod arrow;
 pub mod sql;
 
 #[cfg(test)]
