+++ crates/proof-of-sql/src/sql/proof/final_round_builder_test.rs
@@ -8,13 +8,6 @@ use crate::{
     },
     sql::proof::SumcheckSubpolynomialType,
 };
-use alloc::sync::Arc;
-#[cfg(feature = "arrow")]
-use arrow::{
-    array::Int64Array,
-    datatypes::{Field, Schema},
-    record_batch::RecordBatch,
-};
 use curve25519_dalek::RistrettoPoint;
 use num_traits::{One, Zero};
 
@@ -131,39 +124,6 @@ fn we_can_form_an_aggregated_sumcheck_polynomial() {
     let eval = poly.evaluate(&random_point);
     let expected_eval = expected_poly.evaluate(&random_point);
     assert_eq!(eval, expected_eval);
-}
-
-#[cfg(feature = "arrow")]
-#[test]
-fn we_can_form_the_provable_query_result() {
-    let col1: Column<Curve25519Scalar> = Column::BigInt(&[11_i64, 12]);
-    let col2: Column<Curve25519Scalar> = Column::BigInt(&[-3_i64, -4]);
-    let res = ProvableQueryResult::new(2, &[col1, col2]);
-
-    let column_fields = vec![
-        ColumnField::new("a".parse().unwrap(), ColumnType::BigInt),
-        ColumnField::new("b".parse().unwrap(), ColumnType::BigInt),
-    ];
-    let res = RecordBatch::try_from(
-        res.to_owned_table::<Curve25519Scalar>(&column_fields)
-            .unwrap(),
-    )
-    .unwrap();
-    let column_fields: Vec<Field> = column_fields
-        .iter()
-        .map(core::convert::Into::into)
-        .collect();
-    let schema = Arc::new(Schema::new(column_fields));
-
-    let expected_res = RecordBatch::try_new(
-        schema,
-        vec![
-            Arc::new(Int64Array::from(vec![11, 12])),
-            Arc::new(Int64Array::from(vec![-3, -4])),
-        ],
-    )
-    .unwrap();
-    assert_eq!(res, expected_res);
 }
 
 #[test]
